Refactor Navbar types imports and extract data fetching function
Refactored the import paths of the Navbar component types for a cleaner structure, improving readability and maintenance. The types file is now renamed to index.d.ts. Extracted the code block that fetches Navbar data into a separate function (`getNavbarData`). This provides clearer code organization and makes the main component (`NavbarMain.tsx`) cleaner.
