+++ frontend-next-migration/src/widgets/Navbar/model/types/index.d.ts
@@ -77,3 +77,5 @@ export type NavbarBuild = {
     menu: NavbarMenu;
     namedMenu: NamedMenu
 }
+
+export type NavBarType = "Default" | "Clan" | "TeachingPackage" | "GameArt";
